Hello! I'm Nobody! Who are you?
I work across forms and disciplines, which is an uncomfortable place to be, since most of us are specialists. But then I've tried "comfortable" and it doesn't suit me.

I'm here because I want to take my digital poetry to the next level in terms of visual meaning. So far, it's been based on the page, just like analogue poetry. You can get as far as "Un coup de dés ..." or a kind of caligramme ( which is great!) but not a lot further. After all, these forms go back at least to the century before last.